Build fix + seperate files
